+++ pom.xml
@@ -27,7 +27,9 @@
   </developers>
   <modules>
     <module>willow-sshagent-auth</module>
+    <module>willow-download</module>
     <module>willow-messages</module>
+    <module>willow-process-control</module>
     <module>willow-deployer</module>
     <module>willow-utils</module>
     <module>willow-simulate</module>
