+++ willow-deployer/src/main/java/com/nitorcreations/willow/deployer/Main.java
@@ -1,28 +1,28 @@
 package com.nitorcreations.willow.deployer;
 
-import static com.nitorcreations.willow.deployer.PropertyKeys.PROPERTY_KEY_DEPLOYER_LAUNCH_INDEX;
-import static com.nitorcreations.willow.deployer.PropertyKeys.PROPERTY_KEY_DEPLOYER_NAME;
-import static com.nitorcreations.willow.deployer.PropertyKeys.PROPERTY_KEY_DOWNLOAD_DIRECTORY;
-import static com.nitorcreations.willow.deployer.PropertyKeys.PROPERTY_KEY_DOWNLOAD_RETRIES;
-import static com.nitorcreations.willow.deployer.PropertyKeys.PROPERTY_KEY_LAUNCH_URLS;
-import static com.nitorcreations.willow.deployer.PropertyKeys.PROPERTY_KEY_PREFIX_DOWNLOAD_URL;
-import static com.nitorcreations.willow.deployer.PropertyKeys.PROPERTY_KEY_PREFIX_LAUNCH;
-import static com.nitorcreations.willow.deployer.PropertyKeys.PROPERTY_KEY_PREFIX_POST_DOWNLOAD;
-import static com.nitorcreations.willow.deployer.PropertyKeys.PROPERTY_KEY_PREFIX_POST_START;
-import static com.nitorcreations.willow.deployer.PropertyKeys.PROPERTY_KEY_PREFIX_POST_STOP;
-import static com.nitorcreations.willow.deployer.PropertyKeys.PROPERTY_KEY_PREFIX_POST_STOP_OLD;
-import static com.nitorcreations.willow.deployer.PropertyKeys.PROPERTY_KEY_PREFIX_PRE_DOWNLOAD;
-import static com.nitorcreations.willow.deployer.PropertyKeys.PROPERTY_KEY_PREFIX_PRE_START;
-import static com.nitorcreations.willow.deployer.PropertyKeys.PROPERTY_KEY_PREFIX_SHUTDOWN;
-import static com.nitorcreations.willow.deployer.PropertyKeys.PROPERTY_KEY_PREFIX_STATISTICS;
-import static com.nitorcreations.willow.deployer.PropertyKeys.PROPERTY_KEY_PROPERTIES_FILENAME;
-import static com.nitorcreations.willow.deployer.PropertyKeys.PROPERTY_KEY_REMOTE_REPOSITORY;
-import static com.nitorcreations.willow.deployer.PropertyKeys.PROPERTY_KEY_STATISTICS_FLUSHINTERVAL;
-import static com.nitorcreations.willow.deployer.PropertyKeys.PROPERTY_KEY_STATISTICS_URI;
-import static com.nitorcreations.willow.deployer.PropertyKeys.PROPERTY_KEY_SUFFIX_EXTRA_ENV_KEYS;
-import static com.nitorcreations.willow.deployer.PropertyKeys.PROPERTY_KEY_SUFFIX_METHOD;
-import static com.nitorcreations.willow.deployer.PropertyKeys.PROPERTY_KEY_SUFFIX_TIMEOUT;
-import static com.nitorcreations.willow.deployer.PropertyKeys.PROPERTY_KEY_WORKDIR;
+import static com.nitorcreations.willow.properties.PropertyKeys.PROPERTY_KEY_DEPLOYER_LAUNCH_INDEX;
+import static com.nitorcreations.willow.properties.PropertyKeys.PROPERTY_KEY_DEPLOYER_NAME;
+import static com.nitorcreations.willow.properties.PropertyKeys.PROPERTY_KEY_DOWNLOAD_DIRECTORY;
+import static com.nitorcreations.willow.properties.PropertyKeys.PROPERTY_KEY_DOWNLOAD_RETRIES;
+import static com.nitorcreations.willow.properties.PropertyKeys.PROPERTY_KEY_LAUNCH_URLS;
+import static com.nitorcreations.willow.properties.PropertyKeys.PROPERTY_KEY_PREFIX_DOWNLOAD_URL;
+import static com.nitorcreations.willow.properties.PropertyKeys.PROPERTY_KEY_PREFIX_LAUNCH;
+import static com.nitorcreations.willow.properties.PropertyKeys.PROPERTY_KEY_PREFIX_POST_DOWNLOAD;
+import static com.nitorcreations.willow.properties.PropertyKeys.PROPERTY_KEY_PREFIX_POST_START;
+import static com.nitorcreations.willow.properties.PropertyKeys.PROPERTY_KEY_PREFIX_POST_STOP;
+import static com.nitorcreations.willow.properties.PropertyKeys.PROPERTY_KEY_PREFIX_POST_STOP_OLD;
+import static com.nitorcreations.willow.properties.PropertyKeys.PROPERTY_KEY_PREFIX_PRE_DOWNLOAD;
+import static com.nitorcreations.willow.properties.PropertyKeys.PROPERTY_KEY_PREFIX_PRE_START;
+import static com.nitorcreations.willow.properties.PropertyKeys.PROPERTY_KEY_PREFIX_SHUTDOWN;
+import static com.nitorcreations.willow.properties.PropertyKeys.PROPERTY_KEY_PREFIX_STATISTICS;
+import static com.nitorcreations.willow.properties.PropertyKeys.PROPERTY_KEY_PROPERTIES_FILENAME;
+import static com.nitorcreations.willow.properties.PropertyKeys.PROPERTY_KEY_REMOTE_REPOSITORY;
+import static com.nitorcreations.willow.properties.PropertyKeys.PROPERTY_KEY_STATISTICS_FLUSHINTERVAL;
+import static com.nitorcreations.willow.properties.PropertyKeys.PROPERTY_KEY_STATISTICS_URI;
+import static com.nitorcreations.willow.properties.PropertyKeys.PROPERTY_KEY_SUFFIX_EXTRA_ENV_KEYS;
+import static com.nitorcreations.willow.properties.PropertyKeys.PROPERTY_KEY_SUFFIX_METHOD;
+import static com.nitorcreations.willow.properties.PropertyKeys.PROPERTY_KEY_SUFFIX_TIMEOUT;
+import static com.nitorcreations.willow.properties.PropertyKeys.PROPERTY_KEY_WORKDIR;
 
 import java.io.File;
 import java.io.FileOutputStream;
@@ -63,7 +63,6 @@ import org.hyperic.sigar.ptql.ProcessQueryFactory;
 import com.nitorcreations.willow.deployer.launch.LaunchCallback;
 import com.nitorcreations.willow.deployer.launch.LaunchMethod;
 import com.nitorcreations.willow.deployer.statistics.StatisticsSender;
-import com.nitorcreations.willow.download.PreLaunchDownloadAndExtract;
 import com.nitorcreations.willow.messages.WebSocketTransmitter;
 import com.nitorcreations.willow.messages.event.DeployerStartEvent;
 import com.nitorcreations.willow.messages.event.DeployerStopEvent;
