+++ willow-process-control/src/main/java/com/nitorcreations/willow/deployer/launch/DependencyLauncher.java
@@ -1,7 +1,7 @@
 package com.nitorcreations.willow.deployer.launch;
 
-import static com.nitorcreations.willow.deployer.PropertyKeys.PROPERTY_KEY_SUFFIX_ARTIFACT;
-import static com.nitorcreations.willow.deployer.PropertyKeys.PROPERTY_KEY_SUFFIX_RESOLVE_TRANSITIVE;
+import static com.nitorcreations.willow.properties.PropertyKeys.PROPERTY_KEY_SUFFIX_ARTIFACT;
+import static com.nitorcreations.willow.properties.PropertyKeys.PROPERTY_KEY_SUFFIX_RESOLVE_TRANSITIVE;
 
 import java.io.File;
