+++ willow-process-control/src/main/java/com/nitorcreations/willow/deployer/launch/NativeLauncher.java
@@ -1,7 +1,7 @@
 package com.nitorcreations.willow.deployer.launch;
 
-import static com.nitorcreations.willow.deployer.PropertyKeys.PROPERTY_KEY_SUFFIX_ARGS;
-import static com.nitorcreations.willow.deployer.PropertyKeys.PROPERTY_KEY_SUFFIX_BINARY;
+import static com.nitorcreations.willow.properties.PropertyKeys.PROPERTY_KEY_SUFFIX_ARGS;
+import static com.nitorcreations.willow.properties.PropertyKeys.PROPERTY_KEY_SUFFIX_BINARY;
 
 import javax.inject.Named;
