+++ willow-utils/src/main/java/com/nitorcreations/willow/properties/PropertyKeys.java
@@ -1,4 +1,4 @@
-package com.nitorcreations.willow.deployer;
+package com.nitorcreations.willow.properties;
 
 @SuppressWarnings("PMD.UnusedModifier")
 public interface PropertyKeys {
